+++ jwt/types.go
@@ -5,7 +5,7 @@ import "github.com/golang-jwt/jwt/v4"
 // TokenInfo token info
 type TokenInfo struct {
 	// UserID 用户ID
-	UserID uint `json:"user_id,omitempty"`
+	UserID int64 `json:"user_id,omitempty"`
 	// Post 职位
 	Post string `json:"post,omitempty"`
 }
